+++ CJBCheatsMenu/ModEntry.cs
@@ -5,6 +5,7 @@ using CJBCheatsMenu.Framework;
 using CJBCheatsMenu.Framework.Models;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
+using StardewModdingAPI.Utilities;
 using StardewValley;
 
 namespace CJBCheatsMenu
@@ -15,9 +16,6 @@ namespace CJBCheatsMenu
         /*********
         ** Fields
         *********/
-        /// <summary>The known in-game location.</summary>
-        private Lazy<GameLocation[]> Locations;
-
         /// <summary>The mod settings.</summary>
         private ModConfig Config;
 
@@ -25,7 +23,10 @@ namespace CJBCheatsMenu
         private ModData Warps;
 
         /// <summary>Manages the cheat implementations.</summary>
-        private CheatManager Cheats;
+        private PerScreen<CheatManager> Cheats;
+
+        /// <summary>The known in-game location.</summary>
+        private readonly PerScreen<Lazy<GameLocation[]>> Locations = new PerScreen<Lazy<GameLocation[]>>();
 
 
         /*********
@@ -59,7 +60,7 @@ namespace CJBCheatsMenu
 
             // load cheats
             this.ResetLocationCache();
-            this.Cheats = new CheatManager(this.Config, this.Helper.Reflection, () => this.Locations.Value, this.Warps);
+            this.Cheats = new PerScreen<CheatManager>(() => new CheatManager(this.Config, this.Helper.Reflection, () => this.Locations.Value.Value, this.Warps));
 
             // hook events
             helper.Events.Display.Rendered += this.OnRendered;
@@ -85,7 +86,7 @@ namespace CJBCheatsMenu
         private void OnSaveLoaded(object sender, SaveLoadedEventArgs e)
         {
             this.ResetLocationCache();
-            this.Cheats.OnSaveLoaded();
+            this.Cheats.Value.OnSaveLoaded();
         }
 
         /// <summary>Raised after the player returns to the title screen.</summary>
@@ -114,11 +115,11 @@ namespace CJBCheatsMenu
 
             // open menu
             if (e.Button == this.Config.OpenMenuKey)
-                Game1.activeClickableMenu = new CheatsMenu(this.Config.DefaultTab, this.Cheats, this.Monitor);
+                Game1.activeClickableMenu = new CheatsMenu(this.Config.DefaultTab, this.Cheats.Value, this.Monitor);
 
             // handle button if applicable
             else
-                this.Cheats.OnButtonPressed(e);
+                this.Cheats.Value.OnButtonPressed(e);
         }
 
         /// <summary>Raised after the player releases a button on the keyboard, controller, or mouse.</summary>
@@ -126,7 +127,7 @@ namespace CJBCheatsMenu
         /// <param name="e">The event arguments.</param>
         private void OnButtonReleased(object sender, ButtonReleasedEventArgs e)
         {
-            this.Cheats.OnButtonReleased(e);
+            this.Cheats.Value.OnButtonReleased(e);
         }
 
         /// <summary>Raised after the game draws to the sprite patch in a draw tick, just before the final sprite batch is rendered to the screen.</summary>
@@ -137,7 +138,7 @@ namespace CJBCheatsMenu
             if (!Context.IsWorldReady)
                 return;
 
-            this.Cheats.OnRendered();
+            this.Cheats.Value.OnRendered();
         }
 
         /// <summary>Raised after the game state is updated (≈60 times per second).</summary>
@@ -148,7 +149,7 @@ namespace CJBCheatsMenu
             if (!Context.IsWorldReady)
                 return;
 
-            this.Cheats.OnUpdateTicked(e);
+            this.Cheats.Value.OnUpdateTicked(e);
         }
 
         /// <summary>Raised after a game menu is opened, closed, or replaced.</summary>
@@ -160,17 +161,17 @@ namespace CJBCheatsMenu
             if (e.OldMenu is CheatsMenu)
             {
                 this.Helper.WriteConfig(this.Config);
-                this.Cheats.OnOptionsChanged();
+                this.Cheats.Value.OnOptionsChanged();
             }
         }
 
         /// <summary>Reset the cached location list.</summary>
         private void ResetLocationCache()
         {
-            if (this.Locations?.IsValueCreated == false)
+            if (this.Locations.Value?.IsValueCreated == false)
                 return;
 
-            this.Locations = new Lazy<GameLocation[]>(
+            this.Locations.Value = new Lazy<GameLocation[]>(
                 () => Context.IsWorldReady
                     ? CommonHelper.GetAllLocations().ToArray()
                     : new GameLocation[0]