+++ CJBShowItemSellPrice/ModEntry.cs
@@ -7,6 +7,7 @@ using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
+using StardewModdingAPI.Utilities;
 using StardewValley;
 using StardewValley.Menus;
 using SObject = StardewValley.Object;
@@ -34,14 +35,14 @@ namespace CJBShowItemSellPrice
         /// <summary>The pixel offset to apply to the tooltip box relative to the cursor position.</summary>
         private readonly Vector2 TooltipOffset = new Vector2(Game1.tileSize / 2);
 
-        /// <summary>The cached toolbar instance.</summary>
-        private Toolbar Toolbar;
-
-        /// <summary>The cached toolbar slots.</summary>
-        private IList<ClickableComponent> ToolbarSlots;
-
         /// <summary>Metadata that isn't available from the game data directly.</summary>
         private DataModel Data;
+
+        /// <summary>The cached toolbar instance.</summary>
+        private readonly PerScreen<Toolbar> Toolbar = new PerScreen<Toolbar>();
+
+        /// <summary>The cached toolbar slots.</summary>
+        private readonly PerScreen<IList<ClickableComponent>> ToolbarSlots = new PerScreen<IList<ClickableComponent>>();
 
 
         /*********
@@ -78,15 +79,15 @@ namespace CJBShowItemSellPrice
             {
                 if (Context.IsPlayerFree)
                 {
-                    this.Toolbar = Game1.onScreenMenus.OfType<Toolbar>().FirstOrDefault();
-                    this.ToolbarSlots = this.Toolbar != null
-                        ? this.Helper.Reflection.GetField<List<ClickableComponent>>(this.Toolbar, "buttons").GetValue()
+                    this.Toolbar.Value = Game1.onScreenMenus.OfType<Toolbar>().FirstOrDefault();
+                    this.ToolbarSlots.Value = this.Toolbar.Value != null
+                        ? this.Helper.Reflection.GetField<List<ClickableComponent>>(this.Toolbar.Value, "buttons").GetValue()
                         : null;
                 }
                 else
                 {
-                    this.Toolbar = null;
-                    this.ToolbarSlots = null;
+                    this.Toolbar.Value = null;
+                    this.ToolbarSlots.Value = null;
                 }
             }
         }
@@ -146,18 +147,18 @@ namespace CJBShowItemSellPrice
         /// <summary>Get the hovered item from the on-screen toolbar.</summary>
         private Item GetItemFromToolbar()
         {
-            if (!Context.IsPlayerFree || this.Toolbar == null || this.ToolbarSlots == null || !Game1.displayHUD)
+            if (!Context.IsPlayerFree || this.Toolbar?.Value == null || this.ToolbarSlots == null || !Game1.displayHUD)
                 return null;
 
             // find hovered slot
             int x = Game1.getMouseX();
             int y = Game1.getMouseY();
-            ClickableComponent hoveredSlot = this.ToolbarSlots.FirstOrDefault(slot => slot.containsPoint(x, y));
+            ClickableComponent hoveredSlot = this.ToolbarSlots.Value.FirstOrDefault(slot => slot.containsPoint(x, y));
             if (hoveredSlot == null)
                 return null;
 
             // get inventory index
-            int index = this.ToolbarSlots.IndexOf(hoveredSlot);
+            int index = this.ToolbarSlots.Value.IndexOf(hoveredSlot);
             if (index < 0 || index > Game1.player.Items.Count - 1)
                 return null;